move ThinFileReader into parquet module
