+++ datafusion/core/src/physical_plan/file_format/mod.rs
@@ -434,35 +434,6 @@ impl From<ObjectMeta> for FileMeta {
     }
 }
 
-pub struct ThinFileReader {
-    reader: Box<dyn AsyncFileReader + Send>,
-}
-
-impl AsyncFileReader for ThinFileReader {
-    fn get_bytes(
-        &mut self,
-        range: Range<usize>,
-    ) -> BoxFuture<'_, ::parquet::errors::Result<Bytes>> {
-        self.reader.get_bytes(range)
-    }
-
-    fn get_byte_ranges(
-        &mut self,
-        ranges: Vec<Range<usize>>,
-    ) -> BoxFuture<'_, ::parquet::errors::Result<Vec<Bytes>>>
-    where
-        Self: Send,
-    {
-        self.reader.get_byte_ranges(ranges)
-    }
-
-    fn get_metadata(
-        &mut self,
-    ) -> BoxFuture<'_, ::parquet::errors::Result<Arc<ParquetMetaData>>> {
-        self.reader.get_metadata()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use crate::{
